# services/idun_agent_web/schema/manager-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Idun Agent Manager API",
    "description": "Idun service for managing AI agents",
    "version": "0.5.3"
  },
  "paths": {
...
    "/api/v1/memory/check-connection": {
      "post": {
        "tags": [
          "Memory"
        ],
        "summary": "Check memory store connectivity",
        "description": "Check connectivity to a memory store before saving.",
        "operationId": "check_memory_connection_api_v1_memory_check_connection_post",
        "parameters": [
          {
            "name": "x-project-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          },
          {
            "name": "x-workspace-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/": {
      "post": {
        "tags": [
          "Memory"
        ],
        "summary": "Create managed memory config",
        "description": "Create a new managed memory configuration.",
        "operationId": "create_memory_api_v1_memory__post",
        "parameters": [
          {
            "name": "x-project-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          },
          {
            "name": "x-workspace-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagedMemoryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedMemoryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Memory"
        ],
        "summary": "List managed memory configs",
        "description": "List managed memory configurations with pagination.",
        "operationId": "list_memories_api_v1_memory__get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "agent_framework",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AgentFramework"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Framework"
            }
          },
          {
            "name": "x-project-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          },
          {
            "name": "x-workspace-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManagedMemoryRead"
                  },
                  "title": "Response List Memories Api V1 Memory  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/{id}": {
      "get": {
        "tags": [
          "Memory"
        ],
        "summary": "Get managed memory config by ID",
        "description": "Get a managed memory configuration by ID.",
        "operationId": "get_memory_api_v1_memory__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "x-project-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          },
          {
            "name": "x-workspace-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedMemoryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Memory"
        ],
        "summary": "Delete managed memory config",
        "description": "Delete a managed memory configuration permanently.",
        "operationId": "delete_memory_api_v1_memory__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "x-project-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          },
          {
            "name": "x-workspace-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Memory"
        ],
        "summary": "Update managed memory config",
        "description": "Update a memory configuration.",
        "operationId": "patch_memory_api_v1_memory__id__patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "x-project-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          },
          {
            "name": "x-workspace-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagedMemoryPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedMemoryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdkDatabaseSessionConfig": {
        "properties": {
          "type": {
            "type": "string",
            "const": "database",
            "title": "Type",
            "default": "database"
          },
          "db_url": {
            "type": "string",
            "title": "Db Url",
            "description": "Database URL (e.g. postgresql+psycopg://...)"
          }
        },
        "type": "object",
        "required": [
          "db_url"
        ],
        "title": "AdkDatabaseSessionConfig",
        "description": "Configuration for Database Session Service."
      },
...
      "AdkInMemorySessionConfig": {
        "properties": {
          "type": {
            "type": "string",
            "const": "in_memory",
            "title": "Type",
            "default": "in_memory"
          }
        },
        "type": "object",
        "title": "AdkInMemorySessionConfig",
        "description": "Configuration for In-Memory Session Service."
      },
...
      "AdkVertexAiSessionConfig": {
        "properties": {
          "type": {
            "type": "string",
            "const": "vertex_ai",
            "title": "Type",
            "default": "vertex_ai"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id",
            "description": "Google Cloud Project ID"
          },
          "location": {
            "type": "string",
            "title": "Location",
            "description": "Google Cloud Location (e.g. us-central1)"
          },
          "reasoning_engine_app_name": {
            "type": "string",
            "title": "Reasoning Engine App Name",
            "description": "Reasoning Engine Application Name or ID"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "location",
          "reasoning_engine_app_name"
        ],
        "title": "AdkVertexAiSessionConfig",
        "description": "Configuration for Vertex AI Session Service."
      },
...
      "AgentFramework": {
        "type": "string",
        "enum": [
          "LANGGRAPH",
          "ADK",
          "CREWAI",
          "HAYSTACK",
          "CUSTOM",
          "TRANSLATION_AGENT",
          "CORRECTION_AGENT",
          "DEEP_RESEARCH_AGENT"
        ],
        "title": "AgentFramework",
        "description": "Supported agent frameworks for engine."
      },
...
      "ConnectionCheckResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message",
          "provider",
          "duration_ms"
        ],
        "title": "ConnectionCheckResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InMemoryCheckpointConfig": {
        "properties": {
          "type": {
            "type": "string",
            "const": "memory",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "InMemoryCheckpointConfig",
        "description": "Configuration for In-Memory checkpointer."
      },
...
      "ManagedMemoryCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "agent_framework": {
            "$ref": "#/components/schemas/AgentFramework",
            "description": "Agent framework"
          },
          "memory": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SqliteCheckpointConfig"
              },
              {
                "$ref": "#/components/schemas/InMemoryCheckpointConfig"
              },
              {
                "$ref": "#/components/schemas/PostgresCheckpointConfig"
              },
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/AdkInMemorySessionConfig"
                  },
                  {
                    "$ref": "#/components/schemas/AdkVertexAiSessionConfig"
                  },
                  {
                    "$ref": "#/components/schemas/AdkDatabaseSessionConfig"
                  }
                ],
                "discriminator": {
                  "propertyName": "type",
                  "mapping": {
                    "database": "#/components/schemas/AdkDatabaseSessionConfig",
                    "in_memory": "#/components/schemas/AdkInMemorySessionConfig",
                    "vertex_ai": "#/components/schemas/AdkVertexAiSessionConfig"
                  }
                }
              }
            ],
            "title": "Memory",
            "description": "Memory (checkpoint) configuration"
          }
        },
        "type": "object",
        "required": [
          "name",
          "agent_framework",
          "memory"
        ],
        "title": "ManagedMemoryCreate",
        "description": "Create managed memory model for requests."
      },
      "ManagedMemoryPatch": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "agent_framework": {
            "$ref": "#/components/schemas/AgentFramework",
            "description": "Agent framework"
          },
          "memory": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SqliteCheckpointConfig"
              },
              {
                "$ref": "#/components/schemas/InMemoryCheckpointConfig"
              },
              {
                "$ref": "#/components/schemas/PostgresCheckpointConfig"
              },
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/AdkInMemorySessionConfig"
                  },
                  {
                    "$ref": "#/components/schemas/AdkVertexAiSessionConfig"
                  },
                  {
                    "$ref": "#/components/schemas/AdkDatabaseSessionConfig"
                  }
                ],
                "discriminator": {
                  "propertyName": "type",
                  "mapping": {
                    "database": "#/components/schemas/AdkDatabaseSessionConfig",
                    "in_memory": "#/components/schemas/AdkInMemorySessionConfig",
                    "vertex_ai": "#/components/schemas/AdkVertexAiSessionConfig"
                  }
                }
              }
            ],
            "title": "Memory",
            "description": "Memory (checkpoint) configuration"
          }
        },
        "type": "object",
        "required": [
          "name",
          "agent_framework",
          "memory"
        ],
        "title": "ManagedMemoryPatch",
        "description": "Full replacement schema for PUT of a managed memory."
      },
      "ManagedMemoryRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "agent_framework": {
            "$ref": "#/components/schemas/AgentFramework",
            "description": "Agent framework"
          },
          "memory": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SqliteCheckpointConfig"
              },
              {
                "$ref": "#/components/schemas/InMemoryCheckpointConfig"
              },
              {
                "$ref": "#/components/schemas/PostgresCheckpointConfig"
              },
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/AdkInMemorySessionConfig"
                  },
                  {
                    "$ref": "#/components/schemas/AdkVertexAiSessionConfig"
                  },
                  {
                    "$ref": "#/components/schemas/AdkDatabaseSessionConfig"
                  }
                ],
                "discriminator": {
                  "propertyName": "type",
                  "mapping": {
                    "database": "#/components/schemas/AdkDatabaseSessionConfig",
                    "in_memory": "#/components/schemas/AdkInMemorySessionConfig",
                    "vertex_ai": "#/components/schemas/AdkVertexAiSessionConfig"
                  }
                }
              }
            ],
            "title": "Memory",
            "description": "Memory (checkpoint) configuration"
          },
          "agent_count": {
            "type": "integer",
            "title": "Agent Count",
            "description": "Number of agents using this memory config",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Last update timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "name",
          "agent_framework",
          "memory",
          "created_at",
          "updated_at"
        ],
        "title": "ManagedMemoryRead",
        "description": "Complete managed memory model for responses."
      },
...
      "PostgresCheckpointConfig": {
        "properties": {
          "type": {
            "type": "string",
            "const": "postgres",
            "title": "Type"
          },
          "db_url": {
            "type": "string",
            "title": "Db Url"
          }
        },
        "type": "object",
        "required": [
          "type",
          "db_url"
        ],
        "title": "PostgresCheckpointConfig",
        "description": "Configuration for Postgres checkpointer."
      },
...
      "SqliteCheckpointConfig": {
        "properties": {
          "type": {
            "type": "string",
            "const": "sqlite",
            "title": "Type"
          },
          "db_url": {
            "type": "string",
            "title": "Db Url"
          }
        },
        "type": "object",
        "required": [
          "type",
          "db_url"
        ],
        "title": "SqliteCheckpointConfig",
        "description": "Configuration for SQLite checkpointer."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}